requirements.txt
asgiref==3.6.0
boto3==1.26.128
botocore==1.29.128
certifi==2022.12.7
channels==4.0.0
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2
django-admin==2.0.2
django-cors-headers==3.14.0
django-environ==0.10.0
django-excel-response2==3.0.5
django-on-heroku==1.1.2
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
excel-base==1.0.4
filelock==3.12.0
gunicorn==20.1.0
isoweek==1.3.3
jmespath==1.0.1
pipenv==2023.4.29
platformdirs==3.5.0
postgres==4.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
psycopg2-pool==1.1
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
s3transfer==0.6.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
virtualenv==20.23.0
virtualenv-clone==0.5.7
whitenoise==6.4.0
xlwt==1.3.0
